Fix centering values for BP and HR
From NELA statistician: "The correct centering is done by subtracting 127 from the recorded and winsorized value. This was so in the original model published in Eugene et al (2018) - see the online supplement in: https://www.sciencedirect.com/science/article/pii/S0007091218305786#appsec4. There was no change in this respect when the coefficients were updated in 2020. Analogously, there was no change in the centering value for "Pulse", which is 91."